move constants to separate src/constants folder
