+++ components.json
@@ -15,7 +15,8 @@
     "utils": "@/lib/utils",
     "ui": "@/components/ui",
     "lib": "@/lib",
-    "hooks": "@/hooks"
+    "hooks": "@/hooks",
+    "constants": "@/constants"
   },
   "iconLibrary": "lucide"
 }
